removed comments, implemented stop, fix from gyroscope, fixed x,y,z in upperCase, final implementation of interval variable
