+++ windows/RNSensorsPackage.cs
@@ -3,6 +3,8 @@ using ReactNative.Modules.Core;
 using System.Collections.Generic;
 using ReactNative.UIManager;
 using System;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
 
 namespace RNSensors
 {
@@ -25,6 +27,26 @@ namespace RNSensors
         IReadOnlyList<IViewManager> IReactPackage.CreateViewManagers(ReactContext reactContext)
         {
             return new List<IViewManager>(0);
+        }
+    }
+
+    public class RNSensorsJsonObject
+    {
+        [JsonProperty("x")]
+        public double X;
+
+        [JsonProperty("y")]
+        public double Y;
+
+        [JsonProperty("z")]
+        public double Z;
+
+        [JsonProperty("timestamp")]
+        public DateTimeOffset Timestamp;
+
+        public JObject ToJObject()
+        {
+            return JObject.Parse(JsonConvert.SerializeObject(this));
         }
     }
 }
